requirements.txt
Flask==0.10.1
Flask-Failsafe==0.2
Flask-SQLAlchemy==2.1
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
m3u8==0.2.8
MarkupSafe==0.23
nested-lookup==0.1.3
python-dateutil==2.4.2
requests==2.10.0
six==1.10.0
SQLAlchemy==1.0.12
Werkzeug==0.10.4
xmltodict==0.10.1
